+++ api/graphql/index.js
@@ -2,7 +2,5 @@ module.exports = {
   createIndex: require('./createIndex'),
   makeGraphql: require('./makeGraphql'),
   createResolvers: require('./createResolvers'),
-  createTypeDefs: require('./createTypeDefs'),
-  getGraphqlSchema: require('./getGraphqlSchema'),
-  getGraphqlProperties: require('./getGraphqlProperties'),
+  tcString: require('./tcString'),
 };
