+++ api/graphql/tcString.js
@@ -1,0 +1,3 @@
+module.exports = (name) => {
+
+}
