+++ api/server/serverIndex.js
@@ -1,57 +1,74 @@
 const fs = require('fs');
 const beautify = require('js-beautify').js;
 const { uppercase } = require('../utils');
+const { tcString } = require('../graphql');
 
-module.exports = ({ destination, logging, name, flavor }) => {
+module.exports = ({ schema, destination, logging, flavor }) => {
+  schema = require(schema); // eslint-disable-line
+
+  const isArray = Array.isArray(schema);
+
+  const names = [];
+  if (isArray) {
+    schema.forEach((s) => names.push(uppercase(s.name)));
+  } else {
+    names.push(schema.name);
+  }
   const modelFolder = `${destination}`;
   const indexFile = `${modelFolder}/index.js`;
   // if (logging) console.log('checking server index ');
   const code = [];
-  const graphql = `
-  // graphql
-  const graphqlHTTP = require('express-graphql');
-  // const { makeExecutableSchema } = require('graphql-tools');
-  // const { typeDefs, resolvers } = require('./graphql');
-  // const schema = makeExecutableSchema({ typeDefs, resolvers });
-  // app.use('/graphql', graphqlHTTP({
-  //   schema,
-  //   graphiql: true
-  // }));
 
-  const { composeWithMongoose } = require('graphql-compose-mongoose/node8');
-  const { schemaComposer } = require('graphql-compose');
-  const customizationOptions = {}; // left it empty for simplicity, described below
-  const ${uppercase(name)} = require('./models/${name}');
-  const ${uppercase(name)}TC = composeWithMongoose(${uppercase(name)}, customizationOptions);
+  const graphql = [
+    `
+    // graphql
+    const graphqlHTTP = require('express-graphql');
+    const { composeWithMongoose } = require('graphql-compose-mongoose/node8');
+    const { schemaComposer } = require('graphql-compose');
+    const customizationOptions = {}; // left it empty for simplicity, described below
+    const { ${names.join(', ')} } = require('./models');
+    `
+  ];
 
-  schemaComposer.Query.addFields({
-    ${name}ById: ${uppercase(name)}TC.getResolver('findById'),
-    ${name}ByIds: ${uppercase(name)}TC.getResolver('findByIds'),
-    ${name}One: ${uppercase(name)}TC.getResolver('findOne'),
-    ${name}Many: ${uppercase(name)}TC.getResolver('findMany'),
-    ${name}Count: ${uppercase(name)}TC.getResolver('count'),
-    ${name}Connection: ${uppercase(name)}TC.getResolver('connection'),
-    ${name}Pagination: ${uppercase(name)}TC.getResolver('pagination'),
+
+  names.forEach((name) => {
+    graphql.push(`const ${uppercase(name)}TC = composeWithMongoose(${uppercase(name)}, customizationOptions);`);
   });
 
-  schemaComposer.Mutation.addFields({
-    ${name}CreateOne: ${uppercase(name)}TC.getResolver('createOne'),
-    ${name}CreateMany: ${uppercase(name)}TC.getResolver('createMany'),
-    ${name}UpdateById: ${uppercase(name)}TC.getResolver('updateById'),
-    ${name}UpdateOne: ${uppercase(name)}TC.getResolver('updateOne'),
-    ${name}UpdateMany: ${uppercase(name)}TC.getResolver('updateMany'),
-    ${name}RemoveById: ${uppercase(name)}TC.getResolver('removeById'),
-    ${name}RemoveOne: ${uppercase(name)}TC.getResolver('removeOne'),
-    ${name}RemoveMany: ${uppercase(name)}TC.getResolver('removeMany'),
+  graphql.push('schemaComposer.Query.addFields({');
+  names.forEach((name) => {
+    name = name.toLowerCase();
+    graphql.push(`${name}ById: ${uppercase(name)}TC.getResolver('findById'),`);
+    graphql.push(`${name}ByIds: ${uppercase(name)}TC.getResolver('findByIds'),`);
+    graphql.push(`${name}One: ${uppercase(name)}TC.getResolver('findOne'),`);
+    graphql.push(`${name}Many: ${uppercase(name)}TC.getResolver('findMany'),`);
+    graphql.push(`${name}Count: ${uppercase(name)}TC.getResolver('count'),`);
+    graphql.push(`${name}Connection: ${uppercase(name)}TC.getResolver('connection'),`);
+    graphql.push(`${name}Pagination: ${uppercase(name)}TC.getResolver('pagination'),`);
   });
+  graphql.push('  });');
 
-  const graphqlSchema = schemaComposer.buildSchema();
+  graphql.push('schemaComposer.Mutation.addFields({');
+  names.forEach((name) => {
+    name = name.toLowerCase();
+    graphql.push(`${name}CreateOne: ${uppercase(name)}TC.getResolver('createOne'),`);
+    graphql.push(`${name}CreateMany: ${uppercase(name)}TC.getResolver('createMany'),`);
+    graphql.push(`${name}UpdateById: ${uppercase(name)}TC.getResolver('updateById'),`);
+    graphql.push(`${name}UpdateOne: ${uppercase(name)}TC.getResolver('updateOne'),`);
+    graphql.push(`${name}UpdateMany: ${uppercase(name)}TC.getResolver('updateMany'),`);
+    graphql.push(`${name}RemoveById: ${uppercase(name)}TC.getResolver('removeById'),`);
+    graphql.push(`${name}RemoveOne: ${uppercase(name)}TC.getResolver('removeOne'),`);
+    graphql.push(`${name}RemoveMany: ${uppercase(name)}TC.getResolver('removeMany'),`);
+  });
+  graphql.push('});');
+  graphql.push('const graphqlSchema = schemaComposer.buildSchema();');
+  graphql.push(`
+    app.use('/graphql', graphqlHTTP({
+      schema: graphqlSchema,
+      graphiql: true
+    }));
+  `);
 
-  app.use('/graphql', graphqlHTTP({
-    schema: graphqlSchema,
-    graphiql: true
-  }));
-  `;
   code.push(`
     const express = require('express');
     const fs = require('fs');
@@ -66,7 +83,7 @@ module.exports = ({ destination, logging, name, flavor }) => {
     const jsonParser = bodyParser.json();
     app.use(cors());
     app.use(jsonParser);
-    ${flavor !== 'graphql' ? '' : graphql}
+    ${flavor !== 'graphql' ? '' : graphql.join('')}
     // app.use((req, res, next) => {
     //   console.log(\`\${req.method} \${req.url}\`);
     //   return next();
@@ -74,8 +91,11 @@ module.exports = ({ destination, logging, name, flavor }) => {
     app.use('/health', (req, res) => res.status(200).json({ healthy: true, time: new Date().getTime() }));
 
     // routes
-    app.use([ require('./router/${name}') ]);
   `);
+  names.forEach((name) => {
+    name = name.toLowerCase();
+    code.push(`app.use([ require('./router/${name}') ]);`);
+  });
   code.push('const indexContent = fs.readFileSync(`${abs}/index.html`).toString().replace("https://petstore.swagger.io/v2/swagger.json", `http://localhost:${port}/swagger.json`);'); // eslint-disable-line
   code.push("app.use('/swagger.json', express.static('./swagger.json'));");
   code.push('app.get("/", (req, res) => res.send(indexContent));');
@@ -87,4 +107,4 @@ module.exports = ({ destination, logging, name, flavor }) => {
   const pretty = beautify(code.join('\n'), { indent_size: 2, space_in_empty_paren: true });
 
   fs.writeFileSync(indexFile, pretty);
-}
+};
