+++ index.js
@@ -116,18 +116,15 @@ const letzGetIt = async () => {
   await makeAPI(args);
   await makeSwaggerModelDefinitions(args);
   await makeRouter(args);
-  // await serverIndex(args);
-  // await userCan(args);
-  // await dockerFile(args);
-  // await writePackageJSON(args);
-  // await writeEslint(args);
-  // await writeGitIgnore(args);
-  // await writeDockerIgnore(args);
-  // await readme(args);
-  // await makeTests(args);
-  // if (args.flavor === 'graphql') {
-  //   await makeGraphql(args);
-  // }
+  await serverIndex(args);
+  await userCan(args);
+  await dockerFile(args);
+  await writePackageJSON(args);
+  await writeEslint(args);
+  await writeGitIgnore(args);
+  await writeDockerIgnore(args);
+  await readme(args);
+  await makeTests(args);
   if (args.logging) console.log('all done 🚀'); // eslint-disable-line
 };
